requirements.txt
aiogram
python-dotenv
aiohttp
pytz
